Update button types
